# openapi.yaml
openapi: 3.0.0
info:
  title: 'Medication REST Api'
  version: 1.0.0
paths:
  /api/clients:
    get:
      tags:
        - Clients
      summary: 'Get list of clients'
      description: 'Returns list of clients'
      operationId: getClientsList
      parameters:
        -
          name: per_page
          in: query
          description: 'Number of items per page'
          required: false
          schema:
            type: integer
            default: 15
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Client' } }
                type: object
    post:
      tags:
        - Clients
      summary: 'Store new client'
      description: 'Creates a new client and returns the client data'
      operationId: storeClient
      requestBody:
        description: 'Client data'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreClientRequest'
      responses:
        '201':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/Client' }
                  message: { type: string, example: 'Client created successfully' }
                type: object
        '422':
          description: 'Validation error'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'The given data was invalid.' }
                  errors: { type: object }
                type: object
  '/api/clients/{id}':
    get:
      tags:
        - Clients
      summary: 'Get client information'
      description: 'Returns client data'
      operationId: getClientById
      parameters:
        -
          name: id
          in: path
          description: 'Client id'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/Client' }
                type: object
        '404':
          description: 'Client not found'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'Client not found' }
                type: object
    put:
      tags:
        - Clients
      summary: 'Update existing client'
      description: 'Updates a client and returns the updated data'
      operationId: updateClient
      parameters:
        -
          name: id
          in: path
          description: 'Client id'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: 'Client data'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateClientRequest'
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/Client' }
                  message: { type: string, example: 'Client updated successfully' }
                type: object
        '404':
          description: 'Client not found'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'Client not found' }
                type: object
        '422':
          description: 'Validation error'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'The given data was invalid.' }
                  errors: { type: object }
                type: object
    delete:
      tags:
        - Clients
      summary: 'Delete existing client'
      description: 'Deletes a client and returns a confirmation message'
      operationId: deleteClient
      parameters:
        -
          name: id
          in: path
          description: 'Client id'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'Client deleted successfully' }
                type: object
        '404':
          description: 'Client not found'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'Client not found' }
                type: object
...
components:
  schemas:
...
    StoreClientRequest:
      title: 'Store Client Request'
      description: 'Request for creating a new client'
      required:
        - name
        - age
        - city
        - region
        - income
        - score
        - pin
        - email
        - phone
      properties:
        name:
          description: 'Client name'
          type: string
          example: 'Petr Pavel'
        age:
          description: 'Client age'
          type: integer
          example: 35
        city:
          description: 'Client city'
          type: string
          example: Prague
        region:
          description: 'Client region code (2 characters)'
          type: string
          example: PR
        income:
          description: 'Client monthly income in USD'
          type: number
          format: float
          example: 1500
        score:
          description: 'Client credit score'
          type: integer
          example: 600
        pin:
          description: 'Client personal identification number (must be unique)'
          type: string
          example: 123-45-6789
        email:
          description: 'Client email address (must be unique)'
          type: string
          format: email
          example: petr.pavel@example.com
        phone:
          description: 'Client phone number'
          type: string
          example: '+420123456789'
      type: object
    UpdateClientRequest:
      title: 'Update Client Request'
      description: 'Request for updating an existing client'
      properties:
        name:
          description: 'Client name'
          type: string
          example: 'Petr Pavel'
        age:
          description: 'Client age'
          type: integer
          example: 35
        city:
          description: 'Client city'
          type: string
          example: Prague
        region:
          description: 'Client region code (2 characters)'
          type: string
          example: PR
        income:
          description: 'Client monthly income in USD'
          type: number
          format: float
          example: 1500
        score:
          description: 'Client credit score'
          type: integer
          example: 600
        pin:
          description: 'Client personal identification number (must be unique)'
          type: string
          example: 123-45-6789
        email:
          description: 'Client email address (must be unique)'
          type: string
          format: email
          example: petr.pavel@example.com
        phone:
          description: 'Client phone number'
          type: string
          example: '+420123456789'
      type: object
    Client:
      title: Client
      description: 'Client model'
      properties:
        id:
          description: 'Client ID'
          type: integer
          format: int64
          example: 1
        name:
          description: 'Client name'
          type: string
          example: 'Petr Pavel'
        age:
          description: 'Client age'
          type: integer
          example: 35
        city:
          description: 'Client city'
          type: string
          example: Prague
        region:
          description: 'Client region code'
          type: string
          example: PR
        income:
          description: 'Client monthly income in USD'
          type: number
          format: float
          example: 1500
        score:
          description: 'Client credit score'
          type: integer
          example: 600
        pin:
          description: 'Client personal identification number'
          type: string
          example: 123-45-6789
        email:
          description: 'Client email address'
          type: string
          format: email
          example: petr.pavel@example.com
        phone:
          description: 'Client phone number'
          type: string
          example: '+420123456789'
        created_at:
          description: 'Creation timestamp'
          type: string
          format: date-time
        updated_at:
          description: 'Last update timestamp'
          type: string
          format: date-time
      type: object
...